From: JEA Wallace [mailto:jeaw100@york.ac.uk] Subject: Another weekend, another submission
This time, it's phase 2 of the Deal 'Em input changes (these work well 
with the default cabinet type, I need to do some conditional input work 
to fix them fully), and also a conversion to the resnet code. 
Unfortunately, this still doesn't fix the apparently needed transparency

(the apostrophe and Zenitone logo on the title screen should have a 
green background, and the black rectangle at the bottom should read 
(C)1987). I also tidied up the memory maps, but have been unable to 
reproduce the assert bug relating to VBLANKs, so I'm unsure as to 
whether I've fixed it or not.