From: JEA Wallace [mailto:jeaw100@york.ac.uk] Subject: Deal 'Em phase 3
Conerted the driver to use an HD6845, as the Hitachi part was on the 
PCB, while also using it to generate the VBLANK, rather than hardcoding 
a value. Added the pulldown resistor to the network, and added all the 
conditional inputs. Please note that on the later Cabinet Types (Ready 
Steady Go-1 onwards), the 10p slot is locked out (location of the true 
register for this is unknown), but the input setting is correct - you 
have to use Coin 2 or higher to get money in.