From: Marc Sandusky Subject: Minor patch submission - verbose build information for development in IDE
I am submitting this minor patch that adds verbose build information to the 
build output when the "VERBOSE" flag is set to "1", rather than the simple 
output normally seen, by adding the "-v" flag to the CFLAGS variable.  This 
is most useful for those that like to use an IDE (Eclipse in my case) for 
source browsing and development.  The IDE usually will parse the build 
information, which is currently not available, to determine include file 
locations, build definitions, etc. and will use that information to provide 
more advanced navigation features.