From: Hugh Allen [mailto:hugh2@bigpond.net.au] Subject: patch for state save in buggybjr
Here's a little patch for initial support of state saving in buggybjr. 
(It'll help with tx1 and buggyboy too, but they are non-working)

All I did was get rid of the "anonymous" (temporary) timer that was 
preventing saving, and replace it with a "permanent" timer. I didn't go 
looking for data which might need to be saved.