probably not too interesting.. just some changes to the 'v25' code in toaplan2.c ..
Batsugun does upload what looks to be the entire program, unencrypted.  If enabled it runs to the point of waiting for interrupts, however, for those it looks like it needs a real V25 core with register banks etc.

The other games using the V25 (Dogyuun, Fix Eight, VFive/GrindStormer) only upload a 'stub' program and are encrypted, maybe they upload the rest later?  Knuckle Bash is ROM based V25 (and encrypted).  Charles is working on decrypting them anyway, but core updates would be handy.