From: Steve Lilley-Hopkins [mailto:minwah1959@hotmail.com] Subject: gtmr (kaneko16.c) update Hi,
Please find attached a diff containing a small update to 'Great 100 Miles Rally', within the kaneko16 driver:

* Added 360degree wheel (dial) input to GTMR, which is available when the Controller dip is set to 'Wheel', and Controls dip is set to '1P Side'. Note when Controls dip is set to 'Both Sides', 270degree wheel (paddle) input is used.

I have tested this in the test screen and in-game, and it works as expected. Also note I made a small comment with regard to the GTMR2 wheel controls, which are noted to be not working - they seem to work fine for me.

Any questions/problems please let me know.

Cheers,

Minwah